+++ dbbs_models/__init__.py
@@ -1,11 +1,7 @@
-__version__ = "1.0.3"
-import os, sys
+__version__ = "1.1.3"
 
 from .granule_cell_models import GranuleCell
 from .stellate_cell_models import StellateCell
 from .basket_cell_models import BasketCell
 from .golgi_cell_models import GolgiCell
 from .purkinje_cell_models import PurkinjeCell
-import arborize
-
-arborize.add_directory(os.path.abspath(os.path.join(os.path.dirname(__file__), "morphologies")))